Sort catalogs and tests
