+++ tests/UseCase/SatCatalogosTest.php
@@ -59,6 +59,12 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame('24', $aduana->id());
     }
 
+    public function testCanObtainExistentBanco(): void
+    {
+        $banco = $this->satCatalogos->bancos()->obtain('002');
+        $this->assertSame('002', $banco->id());
+    }
+
     public function testCanObtainExistentClaveUnidad(): void
     {
         $claveUnidad = $this->satCatalogos->clavesUnidades()->obtain('MTK');
@@ -69,24 +75,6 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
     {
         $claveUnidad = $this->satCatalogos->clavesUnidades40()->obtain('MTK');
         $this->assertSame('MTK', $claveUnidad->id());
-    }
-
-    public function testCanObtainExistentObjetoImpuesto40(): void
-    {
-        $productoServicio = $this->satCatalogos->objetosImpuestos40()->obtain('02');
-        $this->assertSame('02', $productoServicio->id());
-    }
-
-    public function testCanObtainExistentProductoServicio(): void
-    {
-        $productoServicio = $this->satCatalogos->productosServicios()->obtain('10101511');
-        $this->assertSame('10101511', $productoServicio->id());
-    }
-
-    public function testCanObtainExistentProductoServicio40(): void
-    {
-        $productoServicio = $this->satCatalogos->productosServicios40()->obtain('10101511');
-        $this->assertSame('10101511', $productoServicio->id());
     }
 
     public function testCanObtainExistentCodigoPostal(): void
@@ -108,16 +96,24 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame('04510', $productoServicio->codigoPostal());
     }
 
-    public function testCanObtainExistentImpuesto(): void
+    public function testCanObtainExistentEstado(): void
     {
-        $impuesto = $this->satCatalogos->impuestos()->obtain('002');
-        $this->assertSame('002', $impuesto->id());
+        $estado = $this->satCatalogos->estados()->obtain('MOR', 'MEX');
+        $this->assertSame('MOR', $estado->codigo());
+        $this->assertSame('MEX', $estado->pais());
     }
 
-    public function testCanObtainExistentImpuesto40(): void
+    public function testCanObtainExistentEstado40(): void
     {
-        $impuesto = $this->satCatalogos->impuestos40()->obtain('002');
-        $this->assertSame('002', $impuesto->id());
+        $estado = $this->satCatalogos->estados40()->obtain('MOR', 'MEX');
+        $this->assertSame('MOR', $estado->codigo());
+        $this->assertSame('MEX', $estado->pais());
+    }
+
+    public function testCanObtainExistentExportacion40(): void
+    {
+        $metodoDePago = $this->satCatalogos->exportaciones40()->obtain('02');
+        $this->assertSame('02', $metodoDePago->id());
     }
 
     public function testCanObtainExistentFormaDePago(): void
@@ -130,6 +126,18 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
     {
         $formaDePago = $this->satCatalogos->formasDePago40()->obtain('03');
         $this->assertSame('03', $formaDePago->id());
+    }
+
+    public function testCanObtainExistentImpuesto(): void
+    {
+        $impuesto = $this->satCatalogos->impuestos()->obtain('002');
+        $this->assertSame('002', $impuesto->id());
+    }
+
+    public function testCanObtainExistentImpuesto40(): void
+    {
+        $impuesto = $this->satCatalogos->impuestos40()->obtain('002');
+        $this->assertSame('002', $impuesto->id());
     }
 
     public function testCanObtainExistentLocalidad40(): void
@@ -176,66 +184,6 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame('BCN', $localidad->estado());
     }
 
-    public function testCanObtainExistentPais(): void
-    {
-        $pais = $this->satCatalogos->paises()->obtain('MEX');
-        $this->assertSame('MEX', $pais->id());
-    }
-
-    public function testCanObtainExistentPais40(): void
-    {
-        $pais = $this->satCatalogos->paises40()->obtain('MEX');
-        $this->assertSame('MEX', $pais->id());
-    }
-
-    public function testCanObtainExistentPeriodicidades40(): void
-    {
-        $pais = $this->satCatalogos->periodicidades40()->obtain('03');
-        $this->assertSame('03', $pais->id());
-    }
-
-    public function testCanObtainExistentRegimenFiscal(): void
-    {
-        $regimenFiscal = $this->satCatalogos->regimenesFiscales()->obtain('601');
-        $this->assertSame('601', $regimenFiscal->id());
-    }
-
-    public function testCanObtainExistentTipoRelacion(): void
-    {
-        $tipoRelacion = $this->satCatalogos->tiposRelaciones()->obtain('05');
-        $this->assertSame('05', $tipoRelacion->id());
-    }
-
-    public function testCanObtainExistentUsoCfdi(): void
-    {
-        $usoCfdi = $this->satCatalogos->usosCfdi()->obtain('G02');
-        $this->assertSame('G02', $usoCfdi->id());
-    }
-
-    public function testCanObtainExistentTipoRelacion40(): void
-    {
-        $tipoRelacion = $this->satCatalogos->tiposRelaciones40()->obtain('05');
-        $this->assertSame('05', $tipoRelacion->id());
-    }
-
-    public function testCanObtainExistentUsoCfdi40(): void
-    {
-        $usoCfdi = $this->satCatalogos->usosCfdi40()->obtain('G02');
-        $this->assertSame('G02', $usoCfdi->id());
-    }
-
-    public function testCanObtainExistentTipoFactor(): void
-    {
-        $tipoFactor = $this->satCatalogos->tiposFactores()->obtain('Tasa');
-        $this->assertSame('Tasa', $tipoFactor->id());
-    }
-
-    public function testCanObtainExistentTipoFactor40(): void
-    {
-        $tipoFactor = $this->satCatalogos->tiposFactores40()->obtain('Tasa');
-        $this->assertSame('Tasa', $tipoFactor->id());
-    }
-
     public function testCanObtainExistentNumeroPedimentoAduana(): void
     {
         $aduana = '43'; // 43 -> c_Aduanas
@@ -272,26 +220,28 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame(999999, $pedimentoAduanaPatente->cantidad());
     }
 
-    public function testCanFindMatchingReglaTasaCuota(): void
+    public function testCanObtainExistentObjetoImpuesto40(): void
     {
-        $rules = $this->satCatalogos->reglasTasaCuota();
-        $this->assertTrue(
-            $rules->hasMatchingRule($rules::IMPUESTO_IVA, $rules::FACTOR_TASA, $rules::USO_TRASLADO, '0.160000')
-        );
-        $this->assertFalse(
-            $rules->hasMatchingRule($rules::IMPUESTO_IVA, $rules::FACTOR_TASA, $rules::USO_TRASLADO, '0.16')
-        );
+        $productoServicio = $this->satCatalogos->objetosImpuestos40()->obtain('02');
+        $this->assertSame('02', $productoServicio->id());
     }
 
-    public function testCanFindMatchingReglaTasaCuota40(): void
+    public function testCanObtainExistentOrigenRecurso(): void
     {
-        $rules = $this->satCatalogos->reglasTasaCuota40();
-        $this->assertTrue(
-            $rules->hasMatchingRule($rules::IMPUESTO_IVA, $rules::FACTOR_TASA, $rules::USO_TRASLADO, '0.160000')
-        );
-        $this->assertFalse(
-            $rules->hasMatchingRule($rules::IMPUESTO_IVA, $rules::FACTOR_TASA, $rules::USO_TRASLADO, '0.16')
-        );
+        $origenRecurso = $this->satCatalogos->origenesRecursos()->obtain('IP');
+        $this->assertSame('IP', $origenRecurso->id());
+    }
+
+    public function testCanObtainExistentPais(): void
+    {
+        $pais = $this->satCatalogos->paises()->obtain('MEX');
+        $this->assertSame('MEX', $pais->id());
+    }
+
+    public function testCanObtainExistentPais40(): void
+    {
+        $pais = $this->satCatalogos->paises40()->obtain('MEX');
+        $this->assertSame('MEX', $pais->id());
     }
 
     public function testCanObtainExistentPatenteAduanal(): void
@@ -306,6 +256,42 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame('9039', $patenteAduanal->id());
     }
 
+    public function testCanObtainExistentPeriodicidades40(): void
+    {
+        $pais = $this->satCatalogos->periodicidades40()->obtain('03');
+        $this->assertSame('03', $pais->id());
+    }
+
+    public function testCanObtainExistentPeriodicidadPago(): void
+    {
+        $periodicidad = $this->satCatalogos->periodicidadesPagos()->obtain('01');
+        $this->assertSame('01', $periodicidad->id());
+    }
+
+    public function testCanObtainExistentProductoServicio(): void
+    {
+        $productoServicio = $this->satCatalogos->productosServicios()->obtain('10101511');
+        $this->assertSame('10101511', $productoServicio->id());
+    }
+
+    public function testCanObtainExistentProductoServicio40(): void
+    {
+        $productoServicio = $this->satCatalogos->productosServicios40()->obtain('10101511');
+        $this->assertSame('10101511', $productoServicio->id());
+    }
+
+    public function testCanObtainExistentRegimenFiscal(): void
+    {
+        $regimenFiscal = $this->satCatalogos->regimenesFiscales()->obtain('601');
+        $this->assertSame('601', $regimenFiscal->id());
+    }
+
+    public function testCanObtainExistentRiesgoPuesto(): void
+    {
+        $riesgoPuesto = $this->satCatalogos->riesgosPuestos()->obtain('1');
+        $this->assertSame('1', $riesgoPuesto->id());
+    }
+
     public function testCanObtainExistentTipoComprobante(): void
     {
         $tipoComprobante = $this->satCatalogos->tiposComprobantes()->obtain('I');
@@ -316,74 +302,6 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
     {
         $tipoComprobante = $this->satCatalogos->tiposComprobantes40()->obtain('I');
         $this->assertSame('I', $tipoComprobante->id());
-    }
-
-    public function testCanObtainExistentTipoNomina(): void
-    {
-        $tipoNomina = $this->satCatalogos->nominas()->obtain('E');
-        $this->assertSame('E', $tipoNomina->id());
-    }
-
-    public function testCanObtainExistentTipoHora(): void
-    {
-        $tipoHora = $this->satCatalogos->horasExtras()->obtain('01');
-        $this->assertSame('01', $tipoHora->id());
-    }
-
-    public function testCanObtainExistentTipoIncapacidad(): void
-    {
-        $tipoIncapacidad = $this->satCatalogos->incapacidades()->obtain('01');
-        $this->assertSame('01', $tipoIncapacidad->id());
-    }
-
-    public function testCanObtainExistentTipoOtroPago(): void
-    {
-        $tipoOtroPago = $this->satCatalogos->otrosTipoPago()->obtain('001');
-        $this->assertSame('001', $tipoOtroPago->id());
-    }
-
-    public function testCanObtainExistentTipoPercepcion(): void
-    {
-        $tipoPercepcion = $this->satCatalogos->percepciones()->obtain('001');
-        $this->assertSame('001', $tipoPercepcion->id());
-    }
-
-    public function testCanObtainExistentBanco(): void
-    {
-        $banco = $this->satCatalogos->bancos()->obtain('002');
-        $this->assertSame('002', $banco->id());
-    }
-
-    public function testCanObtainExistentEstado(): void
-    {
-        $estado = $this->satCatalogos->estados()->obtain('MOR', 'MEX');
-        $this->assertSame('MOR', $estado->codigo());
-        $this->assertSame('MEX', $estado->pais());
-    }
-
-    public function testCanObtainExistentEstado40(): void
-    {
-        $estado = $this->satCatalogos->estados40()->obtain('MOR', 'MEX');
-        $this->assertSame('MOR', $estado->codigo());
-        $this->assertSame('MEX', $estado->pais());
-    }
-
-    public function testCanObtainExistentExportacion40(): void
-    {
-        $metodoDePago = $this->satCatalogos->exportaciones40()->obtain('02');
-        $this->assertSame('02', $metodoDePago->id());
-    }
-
-    public function testCanObtainExistentPeriodicidadPago(): void
-    {
-        $periodicidad = $this->satCatalogos->periodicidadesPagos()->obtain('01');
-        $this->assertSame('01', $periodicidad->id());
-    }
-
-    public function testCanObtainExistentRiesgoPuesto(): void
-    {
-        $riesgoPuesto = $this->satCatalogos->riesgosPuestos()->obtain('1');
-        $this->assertSame('1', $riesgoPuesto->id());
     }
 
     public function testCanObtainExistentTipoContrato(): void
@@ -398,16 +316,52 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame('001', $tipoDeduccion->id());
     }
 
+    public function testCanObtainExistentTipoFactor(): void
+    {
+        $tipoFactor = $this->satCatalogos->tiposFactores()->obtain('Tasa');
+        $this->assertSame('Tasa', $tipoFactor->id());
+    }
+
+    public function testCanObtainExistentTipoFactor40(): void
+    {
+        $tipoFactor = $this->satCatalogos->tiposFactores40()->obtain('Tasa');
+        $this->assertSame('Tasa', $tipoFactor->id());
+    }
+
+    public function testCanObtainExistentTipoHora(): void
+    {
+        $tipoHora = $this->satCatalogos->horasExtras()->obtain('01');
+        $this->assertSame('01', $tipoHora->id());
+    }
+
+    public function testCanObtainExistentTipoIncapacidad(): void
+    {
+        $tipoIncapacidad = $this->satCatalogos->incapacidades()->obtain('01');
+        $this->assertSame('01', $tipoIncapacidad->id());
+    }
+
     public function testCanObtainExistentTipoJornada(): void
     {
         $tipoJornada = $this->satCatalogos->jornadas()->obtain('01');
         $this->assertSame('01', $tipoJornada->id());
     }
 
-    public function testCanObtainExistentOrigenRecurso(): void
+    public function testCanObtainExistentTipoNomina(): void
     {
-        $origenRecurso = $this->satCatalogos->origenesRecursos()->obtain('IP');
-        $this->assertSame('IP', $origenRecurso->id());
+        $tipoNomina = $this->satCatalogos->nominas()->obtain('E');
+        $this->assertSame('E', $tipoNomina->id());
+    }
+
+    public function testCanObtainExistentTipoOtroPago(): void
+    {
+        $tipoOtroPago = $this->satCatalogos->otrosTipoPago()->obtain('001');
+        $this->assertSame('001', $tipoOtroPago->id());
+    }
+
+    public function testCanObtainExistentTipoPercepcion(): void
+    {
+        $tipoPercepcion = $this->satCatalogos->percepciones()->obtain('001');
+        $this->assertSame('001', $tipoPercepcion->id());
     }
 
     public function testCanObtainExistentTipoRegimen(): void
@@ -416,17 +370,27 @@ final class SatCatalogosTest extends UsingTestingDatabaseTestCase
         $this->assertSame('02', $tipoRegimen->id());
     }
 
-    public function testSearchProductosServicios(): void
+    public function testCanObtainExistentTipoRelacion(): void
     {
-        // seed database has 2 records, real database has more than 2
-        $searchResults = $this->satCatalogos->productosServicios()->searchByText('%cerdo%');
-        $this->assertGreaterThanOrEqual(2, count($searchResults));
+        $tipoRelacion = $this->satCatalogos->tiposRelaciones()->obtain('05');
+        $this->assertSame('05', $tipoRelacion->id());
     }
 
-    public function testSearchProductosServicios40(): void
+    public function testCanObtainExistentTipoRelacion40(): void
     {
-        // seed database has 2 records, real database has more than 2
-        $searchResults = $this->satCatalogos->productosServicios40()->searchByText('%cerdo%');
-        $this->assertGreaterThanOrEqual(2, count($searchResults));
+        $tipoRelacion = $this->satCatalogos->tiposRelaciones40()->obtain('05');
+        $this->assertSame('05', $tipoRelacion->id());
+    }
+
+    public function testCanObtainExistentUsoCfdi(): void
+    {
+        $usoCfdi = $this->satCatalogos->usosCfdi()->obtain('G02');
+        $this->assertSame('G02', $usoCfdi->id());
+    }
+
+    public function testCanObtainExistentUsoCfdi40(): void
+    {
+        $usoCfdi = $this->satCatalogos->usosCfdi40()->obtain('G02');
+        $this->assertSame('G02', $usoCfdi->id());
     }
 }
